requirements.txt
chex>=0.1.0
jax>=0.3.21
jaxlib
numpy
optax>=0.1.3
scipy
wandb>=0.13.3
diffrax
matplotlib
equinox>=0.8.0
snax @ git+https://github.com/dieterichlawson/snax.git
tensorflow>=2.8.0
tensorflow-datasets>=4.5.2
tensorflow-probability>=0.15.0
pytest>=7.0.0
Pillow>=9.0.0
